Alter total_cost field in Order model to PositiveIntegerField and update price field in Product model to PositiveIntegerField

+++ product/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Product(models.Model):
     name = models.CharField(max_length=255, verbose_name="Mahsulot nomi")
     description = models.TextField(blank=True, verbose_name="Ta'rif")
-    price = models.DecimalField(max_digits=12, decimal_places=2, verbose_name="Narx")
+    price = models.PositiveIntegerField(verbose_name="Narx")
     is_active = models.BooleanField(default=True, verbose_name="Aktivmi?")
     created_at = models.DateTimeField(auto_now_add=True, verbose_name="Yaratilgan sana")
     updated_at = models.DateTimeField(auto_now=True, verbose_name="Yangilangan sana")
